perf(flamegraph): don't convert to graphviz data unnecessarily
