+++ packages/pyroscope-flamegraph/src/FlameGraph/FlamegraphRenderer.module.scss
@@ -91,59 +91,3 @@
     text-align: right;
   }
 }
-
-.graphVizPane {
-  display: flex;
-  flex: 1;
-
-  // hacky better than !important
-  &#{&} {
-    margin-right: 0;
-    border: 1px solid var(--ps-ui-border);
-  }
-
-  div[id^='graphviz'] {
-    width: 100%;
-    overflow: hidden;
-
-    :global {
-      .graph > polygon {
-        // graphviz overlay
-        fill: none;
-      }
-
-      .node {
-        polygon {
-          // node box
-          // stroke: var(--ps-fl-toolbar-btn-bg);
-        }
-
-        text[text-anchor='middle'] {
-          // node caption
-          // fill: var(--ps-toolbar-icon-color);
-        }
-      }
-
-      .edge {
-        text[text-anchor='middle'] {
-          // edge caption
-          fill: var(--ps-toolbar-icon-color);
-          // fill: red;
-        }
-
-        a {
-          path {
-            // arrow body
-            // stroke: var(--ps-fl-toolbar-btn-bg);
-          }
-
-          polygon {
-            // arrow head
-            // stroke: var(--ps-fl-toolbar-btn-bg);
-            // fill: var(--ps-fl-toolbar-btn-bg);
-          }
-        }
-      }
-    }
-  }
-}
